requirements.txt
opencv-python
opencv-contrib-python
numpy
pillow
mouse
